backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bcrypt==4.0.1
bitarray==2.6.0
boto3==1.24.89
botocore==1.27.89
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cytoolz==0.12.0
dnspython==2.2.1
ecdsa==0.18.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
Faker==15.1.1
fastapi==0.85.0
frozenlist==1.3.1
h11==0.12.0
hexbytes==0.3.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
loguru==0.6.0
lru-dict==1.1.8
MarkupSafe==2.1.1
motor==3.0.0
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
orjson==3.8.0
packaging==21.3
parsimonious==0.8.1
passlib==1.7.4
Pillow==9.2.0
pluggy==1.0.0
protobuf==3.19.5
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
requests==2.28.1
rfc3986==1.5.0
rlp==2.0.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
starlette-context==0.3.4
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
varint==1.0.2
web3==5.31.1
websockets==9.1
yarl==1.8.1
